gjmgj

+++ JAVA/Calculator.java
@@ -3,15 +3,15 @@ import java.util.*;
 public class Calculator {
     public static void main(String args[]) {
         Scanner sc = new Scanner (System.in);
-
+        
         System.out.print("Enter first number: ");
-        int num1 = sc.nextInt();
+        double num1 = sc.nextInt();
 
         System.out.print("Enter the operator (+,-,*,/,%): ");
         char op = sc.next().charAt(0);
 
         System.out.print("Enter second number: ");
-        int num2 = sc.nextInt();
+        double num2 = sc.nextInt();
 
         switch (op) {
             case '+':
